feat(mango): rework internals to allow to lighten the endpoint
