+++ packages/decoder-mango/src/__tests__/detection.spec.ts
@@ -1,7 +1,7 @@
 import { PublicKey, SystemProgram, TransactionInstruction } from '@solana/web3.js';
 import { describe, expect, it } from 'vitest';
 
-import { isMangoInstruction, parseMangoInstructionTitle } from '../detection';
+import { isMangoInstruction } from '../detection';
 import { ENCODED_INSTRUCTIONS, makeInstruction, MANGO_PROGRAM_IDS } from './fixtures';
 
 describe('isMangoInstruction', () => {
@@ -29,24 +29,5 @@ describe('isMangoInstruction', () => {
             new PublicKey('TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA'),
         );
         expect(isMangoInstruction(ix)).toBe(false);
-    });
-});
-
-describe('parseMangoInstructionTitle', () => {
-    it.each([
-        'Deposit',
-        'Withdraw',
-        'AddToBasket',
-        'PlaceSpotOrder',
-        'CancelSpotOrder',
-        'PlacePerpOrder',
-        'PlacePerpOrder2',
-        'CancelPerpOrder',
-        'AddSpotMarket',
-        'AddPerpMarket',
-        'ChangePerpMarketParams',
-    ] as const)('should parse %s instruction', title => {
-        const ix = makeInstruction(ENCODED_INSTRUCTIONS[title], MANGO_PROGRAM_IDS.mainnet);
-        expect(parseMangoInstructionTitle(ix)).toBe(title);
     });
 });
